Move similarity comparator to own file. Correct overuse of Regex.
